spec-proj/requirements.txt
asgiref==3.5.2
attr==0.3.1
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
distlib==0.3.4
Django==4.0.3
django-bootstrap-modal-forms==2.2.0
django-crispy-forms==1.14.0
django-filter==22.1
django-widget-tweaks==1.4.12
filelock==3.7.1
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.0
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
platformdirs==2.5.2
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
svg.path==6.2
textmagic==2.0.3
twilio==7.10.0
tzdata==2021.5
ug==0.0.2
urllib3==1.26.9
virtualenv==20.15.1
Werkzeug==2.1.2
whitenoise==6.2.0
